+++ examples/tracing.rs
@@ -27,20 +27,27 @@ struct StdoutTracker;
 // to actually handle allocation events.  The interface is straightforward: you're notified when an allocation occurs,
 // and when a deallocation occurs.
 impl AllocationTracker for StdoutTracker {
-    fn allocated(&self, addr: usize, size: usize, group_id: AllocationGroupId) {
+    fn allocated(
+        &self,
+        addr: usize,
+        object_size: usize,
+        wrapped_size: usize,
+        group_id: AllocationGroupId,
+    ) {
         // Allocations have all the pertinent information upfront, which you may or may not want to store for further
         // analysis. Notably, deallocations also know how large they are, and what group ID they came from, so you
         // typically don't have to store much data for correlating deallocations with their original allocation.
         println!(
-            "allocation -> addr=0x{:0x} size={} group_id={:?}",
-            addr, size, group_id
+            "allocation -> addr=0x{:0x} object_size={} wrapped_size={} group_id={:?}",
+            addr, object_size, wrapped_size, group_id
         );
     }
 
     fn deallocated(
         &self,
         addr: usize,
-        size: usize,
+        object_size: usize,
+        wrapped_size: usize,
         source_group_id: AllocationGroupId,
         current_group_id: AllocationGroupId,
     ) {
@@ -50,8 +57,8 @@ impl AllocationTracker for StdoutTracker {
         // This can be useful beyond just the obvious "track how many current bytes are allocated by the group", instead
         // going further to see the chain of where allocations end up, and so on.
         println!(
-            "deallocation -> addr=0x{:0x} size={} source_group_id={:?} current_group_id={:?}",
-            addr, size, source_group_id, current_group_id
+            "deallocation -> addr=0x{:0x} object_size={} wrapped_size={} source_group_id={:?} current_group_id={:?}",
+            addr, object_size, wrapped_size, source_group_id, current_group_id
         );
     }
 }